requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.15
pypdf==3.17.1
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.2
watchdog==3.0.0
pdfminer.six==20221105